:recycle: #15 main.goに書かれているコードをinfrastructureに移動
